renaming, split maps for author and publisher
