+++ sling-react-js/src/main/jcr_root/apps/aem-poc/aem-poc-page/aem-poc-page.html
@@ -7,9 +7,7 @@
         <sly data-sly-include="head.html" />
     </head>
     <body data-sly-use.model="com.cathcart93.sling.components.models.aempoc.AemReactPageModel">
-        <!--This block should be hidden in preview-->
 
-        <!--This block should be shown in preview-->
         <div data-sly-test="${!wcmmode.disabled}"
              id="react-root"
              data-author-component="true"
