+++ sling-react-js/src/main/js/aem-poc/components/_componentsMap/serverComponentsMap.js
@@ -1,0 +1,17 @@
+import Parsys from '../_core/parsys';
+import Heading from '../heading/heading';
+import ImageGallery from '../imageGallery/imageGallery';
+import Accordeon from '../accordeon/accordeon';
+import NarrowBanner from '../narrowBanner/narrowBanner';
+import Navigation from '../menu/menu';
+import Page from '../page/page';
+
+export default {
+    Heading,
+    Parsys,
+    ImageGallery,
+    NarrowBanner,
+    Navigation,
+    Page,
+    Accordeon
+}
