+++ sling-react-js/src/main/js/aem-poc/components/_core/aem-publish-component.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import TreeContainer from './TreeContainer'
-import Components from './index.js'
+import Components from '../_componentsMap/serverComponentsMap.js'
 
 const PublishRootComponent = ({config}) => <TreeContainer components={Components} tree={config}/>;
