+++ sling-react-js/src/main/js/aem-poc/components/_core/parsys.js
