+++ sling-react-js/src/main/js/aem-poc/components/heading/heading.js
