--- sling-react-js/src/main/js/aem-poc/components/image-gallery-mock.js
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export default ({totalSlides}) => <h1>React Image Gallery with {totalSlides} should be here</h1>
