--- sling-react-js/src/main/js/aem-poc/components/index.js
@@ -1,20 +1,0 @@
-import Heading from './heading';
-import Parsys from './parsys';
-import ImageGallery from './imageGallery/imageGallery';
-import Accordeon from './accordeon/accordeon';
-import NarrowBanner from './narrowBanner/narrowBanner';
-import Navigation from './menu/menu';
-import Page from './page';
-import AuthorWrapper from './authorWrapper';
-import './main.less';
-
-export default {
-    AuthorWrapper,
-    Heading,
-    Parsys,
-    ImageGallery,
-    NarrowBanner,
-    Navigation,
-    Page,
-    Accordeon
-}
