+++ sling-react-js/src/main/js/aem-poc/components/page/page.less
