+++ sling-react-js/src/main/js/aem-poc/entries/server.aem-poc.js
@@ -1,9 +1,8 @@
 import React from 'react'
 import ReactDOMServer from 'react-dom/server';
-import AemPublishRootComponent from './../components/aem-publish-component'
+import AemPublishRootComponent from '../components/_core/aem-publish-component'
 
 const render = (conf) => {
-    // console.log(conf);
     return ReactDOMServer.renderToString(<AemPublishRootComponent config={conf}/>);
 };
